Subhash - PHOENIX-122 : Migrate Property Tax Information System to Phoenix

+++ egov/egov-demand/src/main/java/org/egov/demand/interfaces/Billable.java
@@ -73,7 +73,7 @@ public interface Billable {
 
 	Date getBillLastDueDate();
 
-	Integer getBoundaryNum();
+	Long getBoundaryNum();
 
 	String getBoundaryType();
